Rename `ssa_type` to `lir_type`
