+++ src/lir/expressions.rs
@@ -2,7 +2,7 @@
 
 use crate::codegen;
 use crate::sema::ast::{FormatArg, StringLocation};
-use crate::lir::ssa_type::Type;
+use crate::lir::lir_type::Type;
 use num_bigint::BigInt;
 use solang_parser::pt::Loc;
 use std::fmt;
